+++ geoarrow-pyarrow/tests/test_dataset.py
@@ -6,35 +6,35 @@ import pyarrow.parquet as pq
 import pytest
 
 import geoarrow.pyarrow as ga
-from geoarrow.pyarrow._dataset import GeoDataset, ParquetRowGroupGeoDataset
+import geoarrow.pyarrow.dataset as gads
 
 
 def test_geodataset_column_name_guessing():
     table = pa.table([ga.array(["POINT (0.5 1.5)"])], ["geometry"])
-    geods = ga.dataset(table)
+    geods = gads.dataset(table)
     assert geods.geometry_columns == ("geometry",)
 
 
 def test_geodataset_column_type_guessing():
     # Already a geoarrow type
     table = pa.table([ga.array(["POINT (0.5 1.5)"])], ["geometry"])
-    geods = ga.dataset(table, geometry_columns=["geometry"])
+    geods = gads.dataset(table, geometry_columns=["geometry"])
     assert geods.geometry_types == (ga.wkt(),)
 
     # utf8 maps to wkt
     table = pa.table([ga.array(["POINT (0.5 1.5)"]).storage], ["geometry"])
-    geods = ga.dataset(table, geometry_columns=["geometry"])
+    geods = gads.dataset(table, geometry_columns=["geometry"])
     assert geods.geometry_types == (ga.wkt(),)
 
     # binary maps to wkb
     table = pa.table([ga.as_wkb(["POINT (0.5 1.5)"]).storage], ["geometry"])
-    geods = ga.dataset(table, geometry_columns=["geometry"])
+    geods = gads.dataset(table, geometry_columns=["geometry"])
     assert geods.geometry_types == (ga.wkb(),)
 
     # Error for other types
     with pytest.raises(TypeError):
         table = pa.table([[123]], ["geometry"])
-        geods = ga.dataset(table, geometry_columns=["geometry"])
+        geods = gads.dataset(table, geometry_columns=["geometry"])
         geods.geometry_types
 
 
@@ -42,12 +42,12 @@ def test_geodataset_in_memory():
     table1 = pa.table([ga.array(["POINT (0.5 1.5)"])], ["geometry"])
     table2 = pa.table([ga.array(["POINT (2.5 3.5)"])], ["geometry"])
 
-    geods = ga.dataset([table1, table2])
+    geods = gads.dataset([table1, table2])
     assert isinstance(geods._parent, ds.InMemoryDataset)
     assert len(list(geods._parent.get_fragments())) == 2
 
     filtered1 = geods.filter_fragments("POLYGON ((2 3, 3 3, 3 4, 2 4, 2 3))")
-    assert isinstance(filtered1, GeoDataset)
+    assert isinstance(filtered1, gads.GeoDataset)
     assert filtered1.to_table().num_rows == 1
     assert filtered1._index.column("_fragment_index") == pa.chunked_array([[0]])
     assert filtered1._index.column("geometry") == geods._index.column("geometry").take(
@@ -59,13 +59,13 @@ def test_geodataset_in_memory():
     assert filtered0.to_table().num_rows == 0
 
     with pytest.raises(TypeError):
-        ga.dataset([table1], use_row_groups=True)
+        gads.dataset([table1], use_row_groups=True)
 
 
 def test_geodataset_in_memory_guessed_type():
     table1 = pa.table([ga.array(["POINT (0.5 1.5)"]).storage], ["geometry"])
     table2 = pa.table([ga.array(["POINT (2.5 3.5)"]).storage], ["geometry"])
-    geods = ga.dataset([table1, table2], geometry_columns=["geometry"])
+    geods = gads.dataset([table1, table2], geometry_columns=["geometry"])
 
     filtered1 = geods.filter_fragments("POLYGON ((2 3, 3 3, 3 4, 2 4, 2 3))")
     assert filtered1.to_table().num_rows == 1
@@ -81,7 +81,7 @@ def test_geodataset_multiple_geometry_columns():
         ["geometry1", "geometry2"],
     )
 
-    geods = ga.dataset([table1, table2])
+    geods = gads.dataset([table1, table2])
     assert isinstance(geods._parent, ds.InMemoryDataset)
     assert len(list(geods._parent.get_fragments())) == 2
 
@@ -98,7 +98,7 @@ def test_geodataset_parquet():
     with TemporaryDirectory() as td:
         pq.write_table(table1, f"{td}/table1.parquet")
         pq.write_table(table2, f"{td}/table2.parquet")
-        geods = ga.dataset(
+        geods = gads.dataset(
             [f"{td}/table1.parquet", f"{td}/table2.parquet"], use_row_groups=False
         )
 
@@ -113,12 +113,12 @@ def test_geodataset_parquet_rowgroups():
     with TemporaryDirectory() as td:
         pq.write_table(table, f"{td}/table.parquet", row_group_size=1)
 
-        geods = ga.dataset(f"{td}/table.parquet")
-        assert isinstance(geods, ParquetRowGroupGeoDataset)
+        geods = gads.dataset(f"{td}/table.parquet")
+        assert isinstance(geods, gads.ParquetRowGroupGeoDataset)
         assert len(geods.get_fragments()) == 2
 
         filtered1 = geods.filter_fragments("POLYGON ((2 3, 3 3, 3 4, 2 4, 2 3))")
-        assert isinstance(filtered1, ParquetRowGroupGeoDataset)
+        assert isinstance(filtered1, gads.ParquetRowGroupGeoDataset)
         assert filtered1.to_table().num_rows == 1
         assert filtered1._index.column("_fragment_index") == pa.chunked_array([[0]])
         assert filtered1._index.column("geometry") == geods._index.column(
@@ -153,10 +153,10 @@ def test_geodataset_parquet_index_rowgroups():
         )
         pq.write_table(table_both, f"{td}/table_both.parquet", row_group_size=1)
 
-        ds_wkt = ga.dataset(f"{td}/table_wkt.parquet")
-        ds_geoarrow = ga.dataset(f"{td}/table_geoarrow.parquet")
-        ds_geoarrow_nostats = ga.dataset(f"{td}/table_geoarrow_nostats.parquet")
-        ds_both = ga.dataset(f"{td}/table_both.parquet")
+        ds_wkt = gads.dataset(f"{td}/table_wkt.parquet")
+        ds_geoarrow = gads.dataset(f"{td}/table_geoarrow.parquet")
+        ds_geoarrow_nostats = gads.dataset(f"{td}/table_geoarrow_nostats.parquet")
+        ds_both = gads.dataset(f"{td}/table_both.parquet")
 
         index_wkt = ds_wkt.index_fragments()
         index_geoarrow = ds_geoarrow.index_fragments()
@@ -186,7 +186,7 @@ def test_geodataset_parquet_filter_rowgroups_with_stats():
     with TemporaryDirectory() as td:
         pq.write_table(table, f"{td}/table.parquet", row_group_size=1)
 
-        geods = ga.dataset(f"{td}/table.parquet")
+        geods = gads.dataset(f"{td}/table.parquet")
         assert len(geods.get_fragments()) == 2
 
         geods._build_index_using_stats(["geometry"])
@@ -199,19 +199,19 @@ def test_geodataset_parquet_filter_rowgroups_with_stats():
 
 def test_parquet_fields_before():
     schema = pa.schema([pa.field("col1", pa.int32()), pa.field("col2", pa.int32())])
-    fields_before = ParquetRowGroupGeoDataset._count_fields_before(schema)
+    fields_before = gads.ParquetRowGroupGeoDataset._count_fields_before(schema)
     assert fields_before == [(("col1",), 0), (("col2",), 1)]
 
     schema = pa.schema(
         [pa.field("col1", pa.list_(pa.int32())), pa.field("col2", pa.int32())]
     )
-    fields_before = ParquetRowGroupGeoDataset._count_fields_before(schema)
+    fields_before = gads.ParquetRowGroupGeoDataset._count_fields_before(schema)
     assert fields_before == [(("col1",), 0), (("col1", "item"), 0), (("col2",), 1)]
 
     schema = pa.schema(
         [pa.field("col1", ga.linestring()), pa.field("col2", pa.int32())]
     )
-    fields_before = ParquetRowGroupGeoDataset._count_fields_before(schema)
+    fields_before = gads.ParquetRowGroupGeoDataset._count_fields_before(schema)
     assert fields_before == [
         (("col1",), 0),
         (("col1", "vertices"), 0),
